Add IO tracing support
I have added a --trace/-t mode to innodb_space and the correspnding
support code to the library in order to trace all read calls through
Innodb::Cursor. I have modified all classes doing reads to pass in
names for the reads they are doing so that the names can be added
to the traces.

This is a lot messier than I would like it to be, but I wanted to
get the code in right now. This could potentially be made cleaner
in the future.