Differentiate between ALLOCATED and FREE pages
Previously pages were returned in space-page-type-regions and
each_page_type_region purely based on the type noted in the page
stored on disk without regard to the extent descriptor's status
for the page. Now we treat pages which are marked free in the
extent descriptor as "FREE" regardless of their type in the actual
page data (since it may be stale).