Support for reading log records.
Introduce Innodb::LogReader which provides a representation of the
log as a seekable stream of log records. The LogReader should be
initialized with a set of reference coordinates (LSN, LSN offset).

A new command named dump-record is added to innodb_log and can be
used to dump a specific record. It uses the log file start LSN as
a reference coordinate for seeking into the log.